refactor(lists): extract ShareListModal into shared component
Moves the share modal out of routes/lists.rs so both the top-level lists
page and the upcoming list-detail settings drawer can reuse it. Adds a
ShareListSection (body without Modal wrapper) for drawer embedding.
