+++ ultros-frontend/ultros-app/src/components/list/mod.rs
@@ -2,3 +2,4 @@ pub mod auto_mark_purchases;
 pub mod buying_view;
 pub mod list_item_row;
 pub mod list_summary;
+pub mod share_list_modal;
